Move non-general methods out of tools.ts, add doc comments to tools.ts methods, add completion on space
